refactor: rename gateways to interfaces in core
