+++ src/core/customers/use-cases/FindCustomerByCpfUseCase.ts
@@ -1,6 +1,6 @@
 import MissingPropertyError from "../../common/exceptions/MissingPropertyError";
 import ResourceNotFoundError from "../../common/exceptions/ResourceNotFoundError";
-import CustomerGateway from "../../gateways/CustomerGateway";
+import CustomerGateway from "../../interfaces/CustomerGateway";
 import FindCustomerByCpf from "../interfaces/FindCustomerByCpf";
 
 export default class FindCustomerByCpfUseCase implements FindCustomerByCpf {
