+++ src/core/interfaces/CustomerGateway.d.ts
