+++ src/core/interfaces/ProductGateway.d.ts
