+++ src/core/orders/use-cases/AddItemUseCase.ts
@@ -1,6 +1,6 @@
 import ResourceNotFoundError from "../../common/exceptions/ResourceNotFoundError";
-import OrderGateway from "../../gateways/OrderGateway";
-import ProductGateway from "../../gateways/ProductGateway";
+import OrderGateway from "../../interfaces/OrderGateway";
+import ProductGateway from "../../interfaces/ProductGateway";
 import ItemDTO from "../dto/ItemDTO";
 import OrderDTO from "../dto/OrderDTO";
 import Item from "../entities/Item";
