+++ src/core/orders/use-cases/GetOrdersUseCase.ts
@@ -1,4 +1,4 @@
-import OrderGateway from "../../gateways/OrderGateway";
+import OrderGateway from "../../interfaces/OrderGateway";
 import ItemDTO from "../dto/ItemDTO";
 import OrderDTO from "../dto/OrderDTO";
 import Item from "../entities/Item";
