+++ src/core/products/use-cases/GetAllProductUseCase.ts
@@ -1,5 +1,5 @@
 import ProductDTO from "../dto/ProductDTO";
-import ProductGateway from "../../gateways/ProductGateway";
+import ProductGateway from "../../interfaces/ProductGateway";
 import GetAllProducts from "../interfaces/GetAllProducts";
 
 export default class GetAllProductsUseCase implements GetAllProducts {
