+++ src/gateways/OrderGateway.ts
@@ -1,4 +1,4 @@
-import OrderGatewayInterface from "../core/gateways/OrderGateway";
+import OrderGatewayInterface from "../core/interfaces/OrderGateway";
 import ItemDTO from "../core/orders/dto/ItemDTO";
 import OrderDTO from "../core/orders/dto/OrderDTO";
 import { OrderDataSource } from "../interfaces/DataSources";
